Reorganise Workspace projection
