+++ crates/but-graph/src/projection/mod.rs
@@ -9,7 +9,6 @@
 ///
 /// Note that these are always a simplification, degenerating information, while maintaining a link back to the graph.
 mod stack;
-
 pub use stack::{Stack, StackCommit, StackCommitDebugFlags, StackCommitFlags, StackSegment};
 
 pub(crate) mod workspace;
