+++ crates/but-graph/src/projection/workspace/init.rs
@@ -1,471 +1,28 @@
 use std::{
     cell::RefCell,
     collections::{BTreeSet, VecDeque},
-    fmt::Formatter,
 };
 
-use anyhow::Context as _;
-use bstr::{BStr, ByteSlice};
-use but_core::{
-    RefMetadata, ref_metadata,
-    ref_metadata::{
-        StackId,
-        StackKind::{Applied, AppliedAndUnapplied},
-    },
-};
-use gix::reference::Category;
+use anyhow::Context;
+use bstr::ByteSlice;
+use but_core::ref_metadata::StackKind::{Applied, AppliedAndUnapplied};
+use but_core::ref_metadata::{self, StackId};
+
+use gix::refs::Category;
 use itertools::Itertools;
-use petgraph::{Direction, prelude::EdgeRef, visit::NodeRef};
+use petgraph::{
+    Direction,
+    visit::{EdgeRef, NodeRef},
+};
 use tracing::instrument;
 
 use crate::{
-    CommitFlags, CommitIndex, Graph, Segment, SegmentIndex,
-    projection::{Stack, StackCommit, StackCommitFlags, StackSegment, workspace},
-    segment,
+    CommitFlags, Graph, Segment, SegmentIndex,
+    projection::{
+        Stack, StackCommit, StackCommitFlags, StackSegment, TargetCommit, TargetRef, Workspace, WorkspaceKind,
+        workspace::{WorkspaceState, find_segment_owner_indexes_by_refname},
+    },
 };
-
-/// A workspace reference is a list of [Stacks](Stack), with a reference to the underlying [`Graph`].
-#[derive(Clone)]
-pub struct Workspace {
-    /// The underlying graph for providing simplified access to data.
-    pub graph: Graph,
-    /// An ID which uniquely identifies the [graph segment](Segment) that represents the tip of the workspace.
-    pub id: SegmentIndex,
-    /// Specify what kind of workspace this is.
-    pub kind: WorkspaceKind,
-    /// One or more stacks that live in the workspace, in order of parents of the workspace commit if there are more than one.
-    pub stacks: Vec<Stack>,
-    /// The bound can be imagined as the commit from which all other commits in the workspace originate.
-    /// It can also be imagined to be the delimiter at the bottom beyond which nothing belongs to the workspace,
-    /// as antagonist to the first commit in tip of the segment with `id`, serving as first commit that is
-    /// inside the workspace.
-    ///
-    /// As such, it's always the longest path to the first shared commit with the target among
-    /// all of our stacks, or it is the first commit that is shared among all of our stacks in absence of a target.
-    /// One can also think of it as the starting point from which all workspace commits can be reached when
-    /// following all incoming connections and stopping at the tip of the workspace.
-    ///
-    /// It is `None` there is only a single stack and no target, so nothing was integrated.
-    pub lower_bound: Option<gix::ObjectId>,
-    /// If `lower_bound` is set, this is the segment owning the commit.
-    pub lower_bound_segment_id: Option<SegmentIndex>,
-    /// The target, as identified by a remote tracking branch, to integrate workspace stacks into.
-    ///
-    /// If `None`, and if `target_commit` is `None`, this is a local workspace that doesn't know when
-    /// possibly pushed branches are considered integrated. This happens when there is a local branch
-    /// checked out without a remote tracking branch.
-    pub target_ref: Option<TargetRef>,
-    /// A commit reachable by [`Self::target_ref`] which we chose to keep as base. That way we can extend the workspace
-    /// past its computed lower bound.
-    ///
-    /// Indeed, it's valid to not set the reference, and to only set the commit which should act as an integration base.
-    pub target_commit: Option<TargetCommit>,
-    /// The segment index of the extra target as provided for traversal,
-    /// useful for AdHoc workspaces, but generally applicable to all workspaces to keep the lower bound lower than it
-    /// otherwise would be.
-    // TODO: could extra-target and target_commit be one and the same? They kind of are, check usages.
-    //       Probably better to keep the `target_commit`.
-    pub extra_target: Option<SegmentIndex>,
-    /// Read-only workspace metadata with additional information, or `None` if nothing was present.
-    /// If this is `Some()` the `kind` is always [`WorkspaceKind::Managed`]
-    pub metadata: Option<ref_metadata::Workspace>,
-}
-
-/// A copy of all workspace state, to pass it around internally.
-pub(crate) struct WorkspaceState {
-    pub id: SegmentIndex,
-    pub kind: WorkspaceKind,
-    pub stacks: Vec<Stack>,
-    pub lower_bound: Option<gix::ObjectId>,
-    pub lower_bound_segment_id: Option<SegmentIndex>,
-    pub target_ref: Option<TargetRef>,
-    pub target_commit: Option<TargetCommit>,
-    pub extra_target: Option<SegmentIndex>,
-    pub metadata: Option<ref_metadata::Workspace>,
-}
-
-impl Workspace {
-    fn from_state(
-        graph: Graph,
-        WorkspaceState {
-            id,
-            kind,
-            stacks,
-            lower_bound,
-            lower_bound_segment_id,
-            target_ref,
-            target_commit,
-            extra_target,
-            metadata,
-        }: WorkspaceState,
-    ) -> Self {
-        Workspace {
-            graph,
-            id,
-            kind,
-            stacks,
-            lower_bound,
-            lower_bound_segment_id,
-            target_ref,
-            target_commit,
-            extra_target,
-            metadata,
-        }
-    }
-}
-
-pub type CommitOwnerIndexes = (usize, usize, CommitIndex);
-
-/// Utilities
-impl Workspace {
-    /// Return the `(merge-base, target-commit-id)` of the merge-base between the `commit_to_merge`
-    /// and either the [target-branch](Self::target_ref), the [extra-target](Self::extra_target)
-    /// or the [target-commit](Self::target_commit), depending on which is set and encountered
-    /// in this order.
-    /// Return `None` when none of these is set, or if there was no merge-base.
-    ///
-    /// Use this to get the merge-base for test-merges between `commit_to_merge` and the target,
-    /// whose commit is also returned as `target-commit-id`.
-    pub fn merge_base_with_target_branch(
-        &self,
-        commit_to_merge: impl Into<gix::ObjectId>,
-    ) -> Option<(gix::ObjectId, gix::ObjectId)> {
-        let commit_to_merge = commit_to_merge.into();
-        let commit_segment_index = self.graph.node_weights().find_map(|s| {
-            s.commits
-                .first()
-                .is_some_and(|c| c.id == commit_to_merge)
-                .then_some(s.id)
-        })?;
-
-        let target_segment_index = self
-            .target_ref
-            .as_ref()
-            .map(|t| t.segment_index)
-            .or(self.target_commit.as_ref().map(|t| t.segment_index))
-            .or(self.extra_target)?;
-
-        let merge_base_segment_index = self
-            .graph
-            .find_git_merge_base(commit_segment_index, target_segment_index)?;
-
-        self.graph
-            .tip_skip_empty(merge_base_segment_index)
-            .map(|c| c.id)
-            .zip(self.graph.tip_skip_empty(target_segment_index).map(|c| c.id))
-    }
-
-    /// Return `true` if the workspace itself is where `HEAD` is pointing to.
-    /// If `false`, one of the stack-segments is checked out instead.
-    pub fn is_entrypoint(&self) -> bool {
-        self.stacks.iter().all(|s| s.segments.iter().all(|s| !s.is_entrypoint))
-    }
-
-    /// Return an iterator over all commits in the workspace,
-    /// i.e. all commits in all segments in all stacks.
-    ///
-    /// This doesn't include the workspace commit.
-    pub fn commits(&self) -> impl Iterator<Item = &StackCommit> + '_ {
-        self.stacks
-            .iter()
-            .flat_map(|s| s.segments.iter())
-            .flat_map(|s| s.commits.iter())
-    }
-
-    /// Return `true` if the branch with `name` is the workspace target or the targets local tracking branch.
-    pub fn is_branch_the_target_or_its_local_tracking_branch(&self, name: &gix::refs::FullNameRef) -> bool {
-        let Some(t) = self.target_ref.as_ref() else {
-            return false;
-        };
-
-        t.ref_name.as_ref() == name
-            || self
-                .graph
-                .lookup_sibling_segment(t.segment_index)
-                .and_then(|local_tracking_segment| local_tracking_segment.ref_name())
-                .is_some_and(|local_tracking_ref| local_tracking_ref == name)
-    }
-
-    /// Return the `commit` at the tip of the workspace itself, and do so by following empty segments along the
-    /// first parent until the first commit is found.
-    /// This importantly is different from the [`Graph::lookup_entrypoint()`] `commit`, as the entrypoint could be anywhere
-    /// inside the workspace as well.
-    ///
-    /// Note that this commit could also be the base of the workspace, particularly if there is no commits in the workspace.
-    pub fn tip_commit(&self) -> Option<&segment::Commit> {
-        self.graph.tip_skip_empty(self.id)
-    }
-
-    /// Lookup a triple obtained by [`Self::find_owner_indexes_by_commit_id()`] or panic.
-    pub fn lookup_commit(&self, (stack_idx, seg_idx, cidx): CommitOwnerIndexes) -> &StackCommit {
-        &self.stacks[stack_idx].segments[seg_idx].commits[cidx]
-    }
-
-    /// Find a stack with the given `id` or error.
-    pub fn try_find_stack_by_id(&self, id: impl Into<Option<StackId>>) -> anyhow::Result<&Stack> {
-        let id = id.into();
-        self.find_stack_by_id(id)
-            .with_context(|| format!("Couldn't find stack with id {id:?} in workspace"))
-    }
-
-    /// Find a stack with the given `id`.
-    pub fn find_stack_by_id(&self, id: impl Into<Option<StackId>>) -> Option<&Stack> {
-        let id = id.into();
-        self.stacks.iter().find(|s| s.id == id)
-    }
-
-    /// Try to find the `(stack_idx, segment_idx, commit_idx)` to be able to access the commit with `oid` in this workspace
-    /// as `ws.stacks[stack_idx].segments[segment_idx].commits[commit_idx]`.
-    pub fn find_owner_indexes_by_commit_id(&self, oid: impl Into<gix::ObjectId>) -> Option<CommitOwnerIndexes> {
-        let oid = oid.into();
-        self.stacks.iter().enumerate().find_map(|(stack_idx, stack)| {
-            stack.segments.iter().enumerate().find_map(|(seg_idx, seg)| {
-                seg.commits
-                    .iter()
-                    .enumerate()
-                    .find_map(|(cidx, c)| (c.id == oid).then_some((stack_idx, seg_idx, cidx)))
-            })
-        })
-    }
-
-    /// Like [`Self::find_owner_indexes_by_commit_id()`], but returns an error if the commit can't be found.
-    pub fn try_find_owner_indexes_by_commit_id(
-        &self,
-        oid: impl Into<gix::ObjectId>,
-    ) -> anyhow::Result<CommitOwnerIndexes> {
-        let oid = oid.into();
-        self.find_owner_indexes_by_commit_id(oid)
-            .with_context(|| format!("Commit {oid} isn't part of the workspace"))
-    }
-
-    /// Try to find the `(stack_idx, segment_idx)` to be able to access the named segment going by `name`.
-    /// Access the segment as `ws.stacks[stack_idx].segments[segment_idx]`
-    pub fn find_segment_owner_indexes_by_refname(&self, ref_name: &gix::refs::FullNameRef) -> Option<(usize, usize)> {
-        find_segment_owner_indexes_by_refname(&self.stacks, ref_name)
-    }
-
-    /// Like [`Self::find_segment_owner_indexes_by_refname`], but fails with an error.
-    pub fn try_find_segment_owner_indexes_by_refname(
-        &self,
-        name: &gix::refs::FullNameRef,
-    ) -> anyhow::Result<(usize, usize)> {
-        self.find_segment_owner_indexes_by_refname(name).with_context(|| {
-            format!(
-                "Couldn't find any stack that contained the branch named '{}'",
-                name.shorten()
-            )
-        })
-    }
-
-    /// Return `true` if `name` is contained in the workspace as segment.
-    pub fn refname_is_segment(&self, name: &gix::refs::FullNameRef) -> bool {
-        self.find_segment_and_stack_by_refname(name).is_some()
-    }
-
-    /// Return `true` if `name` is in the ancestry of the workspace entrypoint, and is IN the workspace as well.
-    pub fn is_reachable_from_entrypoint(&self, name: &gix::refs::FullNameRef) -> bool {
-        if self.ref_name().filter(|_| self.is_entrypoint()) == Some(name) {
-            return true;
-        }
-        if self.is_entrypoint() {
-            self.refname_is_segment(name)
-        } else {
-            let Some((entrypoint_stack, entrypoint_segment_idx)) = self.stacks.iter().find_map(|stack| {
-                stack
-                    .segments
-                    .iter()
-                    .enumerate()
-                    .find_map(|(idx, segment)| segment.is_entrypoint.then_some((stack, idx)))
-            }) else {
-                return false;
-            };
-            entrypoint_stack
-                .segments
-                .get(entrypoint_segment_idx..)
-                .into_iter()
-                .any(|segments| segments.iter().any(|s| s.ref_name().is_some_and(|rn| rn == name)))
-        }
-    }
-
-    /// Try to find `name` in any named [`StackSegment`] and return it along with the stack containing it.
-    pub fn find_segment_and_stack_by_refname(&self, name: &gix::refs::FullNameRef) -> Option<(&Stack, &StackSegment)> {
-        self.stacks.iter().find_map(|stack| {
-            stack
-                .segments
-                .iter()
-                .find_map(|seg| seg.ref_name().is_some_and(|rn| rn == name).then_some((stack, seg)))
-        })
-    }
-
-    /// Like [`Self::find_segment_and_stack_by_refname`], but fails with an error.
-    pub fn try_find_segment_and_stack_by_refname(
-        &self,
-        name: &gix::refs::FullNameRef,
-    ) -> anyhow::Result<(&Stack, &StackSegment)> {
-        self.find_segment_and_stack_by_refname(name).with_context(|| {
-            format!(
-                "Couldn't find any stack that contained the branch named '{}'",
-                name.shorten()
-            )
-        })
-    }
-}
-
-/// Lifecycle
-impl Workspace {
-    /// Redo the graph traversal with the same settings as before, but use the latest
-    /// data from `repo` and `meta` to do it.
-    /// This is useful to make this instance represent changes to `repo` or `meta`.
-    #[instrument(name = "Workspace::refresh_from_head", level = "debug", skip_all, err(Debug))]
-    pub fn refresh_from_head(&mut self, repo: &gix::Repository, meta: &impl RefMetadata) -> anyhow::Result<()> {
-        let graph = Graph::from_head(repo, meta, self.graph.options.clone())?;
-        *self = graph.into_workspace()?;
-        Ok(())
-    }
-}
-
-/// A classifier for the workspace.
-#[derive(Debug, Clone)]
-pub enum WorkspaceKind {
-    /// The `HEAD` is pointing to a dedicated workspace reference, like `refs/heads/gitbutler/workspace`.
-    /// This also means that we have a workspace commit that `ref_name` points to directly, which is also owned
-    /// exclusively by the underlying segment.
-    Managed {
-        /// The name of the reference pointing to the workspace commit, along with workspace info. Useful for deriving the workspace name.
-        ref_info: crate::RefInfo,
-    },
-    /// Information for when a workspace reference was *possibly* advanced by hand and does not point to a
-    /// managed workspace commit (anymore).
-    /// That workspace commit, may be reachable by following the first parent from the workspace reference.
-    ///
-    /// Note that the stacks that follow *will* be in unusable if the workspace commit is in a segment below,
-    /// but typically is usable if there is just a single real stack, or any amount of virtual stacks below
-    /// (i.e. those that have no commits and are just marked by references).
-    ManagedMissingWorkspaceCommit {
-        /// The name of the reference pointing to the workspace commit. Useful for deriving the workspace name.
-        ref_info: crate::RefInfo,
-    },
-    /// A segment is checked out directly.
-    ///
-    /// It can be inside or outside a workspace.
-    /// If the respective segment is [not named](Workspace::ref_name), this means the `HEAD` id detached.
-    /// The commit that the working tree is at is always implied to be the first commit of the [`StackSegment`]
-    /// at [`Workspace::id`].
-    AdHoc,
-}
-
-impl WorkspaceKind {
-    /// Return `true` if this workspace has a managed reference, meaning we control certain aspects of it
-    /// by means of workspace metadata that is associated with that ref.
-    /// If `false`, we are more conservative and may not support all features.
-    pub fn has_managed_ref(&self) -> bool {
-        matches!(
-            self,
-            WorkspaceKind::Managed { .. } | WorkspaceKind::ManagedMissingWorkspaceCommit { .. }
-        )
-    }
-
-    /// Return `true` if we have a workspace commit, a commit that merges all stacks together.
-    /// Implies `has_managed_ref() == true`.
-    pub fn has_managed_commit(&self) -> bool {
-        matches!(self, WorkspaceKind::Managed { .. })
-    }
-}
-
-impl WorkspaceKind {
-    fn managed(ref_info: &Option<crate::RefInfo>) -> anyhow::Result<Self> {
-        let ref_info = ref_info
-            .clone()
-            .context("BUG: managed workspaces must always be on a named segment")?;
-        Ok(WorkspaceKind::Managed { ref_info })
-    }
-}
-
-/// Information about the target reference.
-#[derive(Debug, Clone)]
-pub struct TargetRef {
-    /// The name of the target branch, i.e. the branch that all [Stacks](Stack) want to get merged into.
-    /// Typically, this is `refs/remotes/origin/main`.
-    pub ref_name: gix::refs::FullName,
-    /// The index to the respective segment in the graph, it's the segment with [`Self::ref_name`] as name.
-    pub segment_index: SegmentIndex,
-    /// The amount of commits that aren't included in any segment in the workspace, they are in its future.
-    pub commits_ahead: usize,
-}
-
-/// Information about the target commit.
-#[derive(Debug, Clone)]
-pub struct TargetCommit {
-    /// The hash of the commit that was once included in the [target ref](TargetRef), and that we remember to expand
-    /// the reach of the workspace.
-    pub commit_id: gix::ObjectId,
-    /// The index to the respective segment in the graph for which [`Self::commit_id`] is the first commit.
-    pub segment_index: SegmentIndex,
-}
-
-impl TargetCommit {
-    /// Find `target_commit_id` in the `graph` and store its segment in this instance, or return `None` if not found.
-    fn from_commit(target_commit_id: gix::ObjectId, graph: &Graph) -> Option<Self> {
-        graph.node_weights().find_map(|s| {
-            s.commits.first().and_then(|c| {
-                (c.id == target_commit_id).then_some(TargetCommit {
-                    commit_id: target_commit_id,
-                    segment_index: s.id,
-                })
-            })
-        })
-    }
-}
-
-impl TargetRef {
-    /// Return `None` if `ref_name` wasn't found as segment in `graph`.
-    /// This can happen if a reference is configured, but not actually present as reference.
-    /// Note that `commits_ahead` isn't set yet, see [`Self::compute_and_set_commits_ahead()`].
-    fn from_ref_name_without_commits_ahead(ref_name: &gix::refs::FullName, graph: &Graph) -> Option<Self> {
-        let target_segment_sidx = graph.inner.node_indices().find_map(|n| {
-            let s = &graph[n];
-            (s.ref_name() == Some(ref_name.as_ref())).then_some(s.id)
-        })?;
-        Some(TargetRef {
-            ref_name: ref_name.to_owned(),
-            segment_index: target_segment_sidx,
-            commits_ahead: 0,
-        })
-    }
-
-    fn compute_and_set_commits_ahead(&mut self, graph: &Graph, lower_bound_segment: Option<SegmentIndex>) {
-        let lower_bound = lower_bound_segment.map(|sidx| (sidx, graph[sidx].generation));
-        self.commits_ahead = 0;
-        Self::visit_upstream_commits(graph, self.segment_index, lower_bound, |s| {
-            self.commits_ahead += s.commits.len();
-        })
-    }
-}
-
-/// Utilities
-impl TargetRef {
-    /// Visit all segments whose commits would be considered 'upstream', or part of the target branch
-    /// whose tip is identified with `target_segment`. The `lower_bound_segment_and_generation` is another way
-    /// to stop the traversal.
-    pub fn visit_upstream_commits(
-        graph: &Graph,
-        target_segment: SegmentIndex,
-        lower_bound_segment_and_generation: Option<(SegmentIndex, usize)>,
-        mut visit: impl FnMut(&Segment),
-    ) {
-        graph.visit_all_segments_including_start_until(target_segment, Direction::Outgoing, |s| {
-            let prune = true;
-            if lower_bound_segment_and_generation.is_some_and(|(lower_bound, lower_bound_generation)| {
-                s.id == lower_bound || s.generation > lower_bound_generation
-            }) || s.commits.iter().any(|c| c.flags.contains(CommitFlags::InWorkspace))
-            {
-                return prune;
-            }
-            visit(s);
-            !prune
-        });
-    }
-}
 
 pub(crate) enum Downgrade {
     /// Allows to turn a workspace above a selection to be downgraded back to the selection if it turns
@@ -489,7 +46,7 @@ impl Graph {
     /// Typically, that's a previous location of the target segment.
     #[instrument(name = "Graph::into_workspace", level = "debug", skip(self), err(Debug))]
     pub fn into_workspace(self) -> anyhow::Result<Workspace> {
-        let state = self.to_workspace_state(workspace::Downgrade::Allow)?;
+        let state = self.to_workspace_state(Downgrade::Allow)?;
         Ok(Workspace::from_state(self, state))
     }
 
@@ -1242,103 +799,5 @@ impl WorkspaceState {
         let first_segment = stack.segments.first_mut().expect("non-empty");
         first_segment.commits.clear();
         first_segment.commits_by_segment.clear();
-    }
-}
-
-fn find_segment_owner_indexes_by_refname(
-    stacks: &[Stack],
-    ref_name: &gix::refs::FullNameRef,
-) -> Option<(usize, usize)> {
-    stacks.iter().enumerate().find_map(|(stack_idx, stack)| {
-        stack.segments.iter().enumerate().find_map(|(seg_idx, seg)| {
-            seg.ref_name()
-                .is_some_and(|rn| rn == ref_name)
-                .then_some((stack_idx, seg_idx))
-        })
-    })
-}
-
-/// Query
-impl Workspace {
-    /// Return `true` if the workspace has workspace metadata associated with it.
-    /// This is relevant when creating references for example.
-    pub fn has_metadata(&self) -> bool {
-        self.metadata.is_some()
-    }
-
-    /// Return the name of the workspace reference by looking our segment up in `graph`.
-    /// Note that for managed workspaces, this can be retrieved via [`WorkspaceKind::Managed`].
-    pub fn ref_name(&self) -> Option<&gix::refs::FullNameRef> {
-        self.graph[self.id].ref_name()
-    }
-
-    /// Like [Self::ref_name()], but returns reference and worktree information instead.
-    pub fn ref_info(&self) -> Option<&crate::RefInfo> {
-        self.graph[self.id].ref_info.as_ref()
-    }
-
-    /// Like [`Self::ref_name()`], but return a generic `<anonymous>` name for unnamed workspaces.
-    pub fn ref_name_display(&self) -> &BStr {
-        self.ref_name().map_or("<anonymous>".into(), |rn| rn.as_bstr())
-    }
-}
-
-/// Debugging
-impl Workspace {
-    /// Produce a distinct and compressed debug string to show at a glance what the workspace is about.
-    pub fn debug_string(&self) -> String {
-        let graph = &self.graph;
-        let (name, sign) = match &self.kind {
-            WorkspaceKind::Managed { ref_info } => (
-                Graph::ref_debug_string(ref_info.ref_name.as_ref(), ref_info.worktree.as_ref()),
-                "🏘️",
-            ),
-            WorkspaceKind::ManagedMissingWorkspaceCommit { ref_info } => (
-                Graph::ref_debug_string(ref_info.ref_name.as_ref(), ref_info.worktree.as_ref()),
-                "🏘️⚠️",
-            ),
-            WorkspaceKind::AdHoc => (
-                graph[self.id].ref_info.as_ref().map_or("DETACHED".into(), |ri| {
-                    Graph::ref_debug_string(ri.ref_name.as_ref(), ri.worktree.as_ref())
-                }),
-                "⌂",
-            ),
-        };
-        let target = self.target_ref.as_ref().map_or_else(
-            || "!".to_string(),
-            |t| {
-                format!(
-                    "{target}{ahead}",
-                    target = t.ref_name,
-                    ahead = if t.commits_ahead == 0 {
-                        "".to_string()
-                    } else {
-                        format!("⇣{}", t.commits_ahead)
-                    }
-                )
-            },
-        );
-        format!(
-            "{meta}{sign}:{id}:{name} <> ✓{target}{bound}",
-            meta = if self.metadata.is_some() { "📕" } else { "" },
-            id = self.id.index(),
-            bound = self
-                .lower_bound
-                .map(|base| format!(" on {}", base.to_hex_with_len(7)))
-                .unwrap_or_default()
-        )
-    }
-}
-
-impl std::fmt::Debug for Workspace {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct(&format!("Workspace({})", self.debug_string()))
-            .field("id", &self.id.index())
-            .field("kind", &self.kind)
-            .field("stacks", &self.stacks)
-            .field("metadata", &self.metadata)
-            .field("target_ref", &self.target_ref)
-            .field("extra_target", &self.extra_target)
-            .finish()
     }
 }
